Removed blocking code (refactoring)
Blocking code is not a best practice when using Spray, Akka and Scala Futures.
Hence, I've refactored the code to remove both the CountDownLatch and the Await
on Future.

Furthermore, messages exchanged by actors should always be *immutable*
(otherwise, thread-safety is not guaranteed). This was not the case before this
refactoring, as the `AggregateActor` was sending a mutable collection back upon
reception of the `Result` message.